// frontend/types/Patient.ts
export type Patient = {
  user_id: string;
  baby_name: string;
  baby_date_of_birth: string;
  parent_name: string;
  parent_mobile: string;
  parent_email: string;
  gender: string;
  blood_group: string;
  address: string;
  city: string;
  state: string;
  pin_code: string;
  first_interaction_date: string;
}